+++ src/Segment/Segment.php
@@ -4,7 +4,8 @@ declare(strict_types=1);
 
 namespace Smeghead\TextLinkEncoder\Segment;
 
-interface Segment {
+interface Segment
+{
     public static function getSearchRegex(): string;
     public function __construct(string $segment);
     public function toHtml(): string;